move helper functions to refinementMappings and remove old arrays
