+++ app/utils/refinementMappings.ts
@@ -1,3 +1,5 @@
+import { RefinementListItem } from "instantsearch.js/es/connectors/refinement-list/connectRefinementList";
+
 // On search results page, not category listings page
 
 // Hardcoded categories in search page sidebar (Can add multiple categories as array)
@@ -108,124 +110,6 @@ export const our415SubcategoryNames = new Set(
   our415SubcategoriesWithDuplicates
 );
 
-// const eligibilitiesToDisplay = [
-//   "African American",
-//   "Asian",
-//   "Latinx",
-//   "Pacific Islander",
-//   "Middle Eastern and North African",
-//   "Native American",
-//   "Filipino/a",
-//   "Preschool",
-//   "Elementary School",
-//   "Middle School Students",
-//   "High School Students",
-//   "College Students",
-//   "City College of SF",
-//   "Low-Income",
-//   "Underinsured",
-//   "Uninsured",
-//   "Benefit Recipients",
-//   "Foster Youth",
-//   "I am Experiencing Homelessness",
-//   "Undocumented",
-//   "CIP (Children of Incarcerated Parents)",
-//   "Trauma Survivors",
-//   "Teen Parents",
-//   "Non-Binary",
-//   "Transgender and Gender Non-Conforming",
-//   "Boys",
-//   "Men",
-//   "Girls",
-//   "Women",
-//   "LGBTQ+",
-//   "Youth",
-//   "Teens",
-//   "Toddler",
-//   "Families with children below 18 years old",
-//   "Pre-Teen",
-//   "Infants",
-//   "All Ages",
-//   "TAY",
-//   "Children",
-//   "Special Needs/Disabilities",
-//   "Mental and Behavioral Health Challenges",
-//   "Deaf or Hard of Hearing",
-//   "Visual Impairment",
-//   "Limited Mobility",
-//   "Spanish",
-//   "Chinese",
-//   "Filipino",
-//   "ESL/ELL (English Language Learner)",
-// ];
-
-// const allEligibilities = [
-//   "API (Asian/Pacific Islander", // Has results if Pacific Islander does not
-//   "African American",
-//   "All Ages",
-//   "Asian",
-//   "Benefit Recipients", // No results
-//   "Boys",
-//   "CIP (Children of Incarcerated Parents)",
-//   "Children", // From DCYF, no results
-//   "Children (0-13 years old)", // Has results if Children does not
-//   "Chinese",
-//   "City College of SF", // From DCYF, no results
-//   "College Students", // Has results, while College does not
-//   "College", // From DCYF, no results
-//   "Deaf or Hard of Hearing", // From DCYF, no results
-//   "Elementary School", // From DCYF, no results
-//   "Elementary School Student", // Has results if Elementary School does not
-//   "ESL/ELL (English Language Learner)",
-//   "Families with children below 18 years old",
-//   "Filipino", // From DCYF, no results
-//   "Filipino/a",
-//   "Foster Youth",
-//   "Girls",
-//   "High School Students", // From DCYF, no results
-//   "High School", // Has results if High School Students does not
-//   "I am Experiencing Homelessness", // From DCYF, no results
-//   "Experiencing Homelessness", // Has results if I am Experiencing Homelessness does not
-//   "Infants", // From DCYF, no results
-//   "Infants (0-2 years old)", // Has results if Infants does not
-//   "Latinx", // From DCYF, no results
-//   "Hispanic/Latinx", // Has results if Latinx does not
-//   "LGBTQ+",
-//   "Limited Mobility", // From DCYF, no results
-//   "Low-Income",
-//   "Mental and Behavioral Health Challenges", // From DCYF, no results
-//   "Men",
-//   "Middle Eastern and North African",
-//   "Middle School Students", // From DCYF, no results
-//   "Middle School", // Has results if Middle School Students does not
-//   "Native American",
-//   "Non-Binary",
-//   "Pacific Islander",
-//   "Preschool", // From DCYF, no results
-//   "Preschool Student", // Has results if Preschool does not
-//   "Pre-Teen", // From DCYF, no results
-//   "Preteen", // Has results if Pre-Teen does not
-//   "Spanish", // From DCYF, no results
-//   "Special Needs/Disabilities",
-//   "TAY", // From DCYF, no results
-//   "Transitional Aged Youth (TAY)", // Has results if TAY does not
-//   "Teen Parent", // Has results if Teen Parents does not
-//   "Teen Parents", // From DCYF, no results
-//   "Teens", // From DCYF, no results
-//   "Teens (13-18 years old)", // Has results if Teens does not
-//   "Toddler", // From DCYF, no results
-//   "Toddler Age", // Has results if Toddler does not
-//   "Transgender and Gender Non-Conforming",
-//   "Trauma Survivors",
-//   "Underinsured",
-//   "Undocumented",
-//   "Uninsured",
-//   "Visual Impairment",
-//   "Women",
-//   "Youth", // From DCYF, no results
-//   "Youth (below 21 years old)", // Has results if Youth does not
-// ];
-
 export const our415EligibilitiesMapping = {
   "African American": ["African American"],
   Asian: ["Asian"],
@@ -285,4 +169,47 @@ export const our415EligibilitiesMapping = {
   Spanish: ["Spanish"],
   Chinese: ["Chinese"],
   "ESL/ELL (English Language Learner)": ["ESL/ELL (English Language Learner)"],
+};
+
+// only return eligibilities that are in the refinement map
+export const filterItemsUsingMapping = (
+  items: RefinementListItem[],
+  mapping: Record<string, string[]>
+): RefinementListItem[] => {
+  return items.filter((item) =>
+    Object.values(mapping).some((apiEligibilities) =>
+      apiEligibilities.includes(item.label)
+    )
+  );
+};
+
+export const transformItemsUsingMapping = (
+  items: RefinementListItem[],
+  mapping: Record<string, string[]>
+): RefinementListItem[] => {
+  return items.map((item) => {
+    // loook through each mapping key to see if the item label is one of its values
+    const mappedEligibility = Object.entries(mapping).find(
+      ([, apiEligibilities]) => apiEligibilities.includes(item.label)
+    );
+    if (mappedEligibility) {
+      const [mappedKey] = mappedEligibility;
+      return { ...item, label: mappedKey }; // same object but with mapped key
+    }
+    return item;
+  });
+};
+
+export const deduplicateItemsByLabel = (
+  items: RefinementListItem[]
+): RefinementListItem[] => {
+  const seen = new Set();
+  return items.filter((item) => {
+    if (seen.has(item.label)) {
+      return false;
+    } else {
+      seen.add(item.label);
+      return true;
+    }
+  });
 };
